Organize initial data files

+++ manage.py
@@ -55,7 +55,12 @@ def cleanup():
                 elif os.path.isdir(full_db_file):
                     rmtree(full_db_file)
                 print 'remove', full_db_file
-    db.drop_all()
+        db.drop_all()
+    else:
+        confirm = raw_input(u'Are you sure to clean up the database? [Y/n]: ')
+        if confirm == u'Y':
+            db.drop_all()
+            print '---> All data are deleted.'
 
 
 @manager.command
